+++ UNITE_Main.X/CommandParser.h
@@ -80,6 +80,6 @@ void ParseByte(uint8_t, CommandByteIndex);
 
 void addValueByte(uint8_t);
 uint16_t convertHexToDecimal(uint8_t *);
-uint16_t convertTime(Unit,Unit);
-void RunCommand(System, Mode, Property, Unit, uint16_t);
+unsigned long convertTime(Unit,Unit);
+void RunCommand(System, Mode, Property, Unit, unsigned long);
